refactor: 로또 번호, 당첨 번호에 대한 새로운 도메인 클래스로 분리
